Add files via upload
For a Hospital Management System (HMS) project using Hibernate, here's a sample structure for the User Menu that includes key functionalities you might need for different types of users such as Admin, Doctor, Nurse, and Patient. This example will focus on the Hibernate-specific actions for persisting and retrieving data from a database.